+++ src/image.rs
@@ -1,4 +1,3 @@
-use std;
 use std::mem::MaybeUninit;
 use std::os::raw::c_int;
 use std::ptr;